tests/requirements.txt
playwright==1.55.0
pytest-playwright==0.7.0
pyyaml
opencv-python
pytest-html
